Moved LockBuilder to Lock.Builder.
Removed unnecesary setters in Lock object.
