+++ lock/src/test/java/com/auth0/lock/BuilderTest.java
@@ -51,7 +51,7 @@ import static org.junit.Assert.assertThat;
 
 @RunWith(RobolectricTestRunner.class)
 @Config(constants = BuildConfig.class, sdk = 18, manifest = Config.NONE)
-public class LockBuilderTest {
+public class BuilderTest {
 
     private static final String CLIENT_ID = "CLIENTID";
     private static final String TENANT = "TENANT";
@@ -62,7 +62,7 @@ public class LockBuilderTest {
     private static final String AUTH0_SUBDOMAIN = "pepe.auth0.com";
     private static final String EU_DOMAIN = "samples.eu.auth0.com";
 
-    private LockBuilder builder;
+    private Lock.Builder builder;
     private Lock lock;
 
     @Mock
@@ -281,7 +281,7 @@ public class LockBuilderTest {
         assertThat(lock.getCredentialStore(), is(instanceOf(NullCredentialStore.class)));
     }
 
-    private LockBuilder basicBuilder() {
+    private Lock.Builder basicBuilder() {
         return builder
                 .clientId(CLIENT_ID)
                 .tenant(TENANT);
